# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FirstSeenIndicatorOut": {
        "properties": {
          "indicator_hash": {
            "title": "Indicator Hash",
            "type": "string"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/FirstSeenTenantObservationOut"
            },
            "title": "Observations",
            "type": "array"
          },
          "tenant_count": {
            "title": "Tenant Count",
            "type": "integer"
          }
        },
        "required": [
          "indicator_hash",
          "tenant_count",
          "observations"
        ],
        "title": "FirstSeenIndicatorOut",
        "type": "object"
      },
      "FirstSeenResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FirstSeenIndicatorOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "FirstSeenResponse",
        "type": "object"
      },
      "FirstSeenTenantObservationOut": {
        "properties": {
          "first_observed_at": {
            "format": "date-time",
            "title": "First Observed At",
            "type": "string"
          },
          "tenant_hash": {
            "title": "Tenant Hash",
            "type": "string"
          }
        },
        "required": [
          "tenant_hash",
          "first_observed_at"
        ],
        "title": "FirstSeenTenantObservationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentTypeBreakdownOut": {
        "properties": {
          "avg_confidence": {
            "title": "Avg Confidence",
            "type": "number"
          },
          "avg_evidence_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Evidence Confidence"
          },
          "evidence_confidence_count": {
            "default": 0,
            "title": "Evidence Confidence Count",
            "type": "integer"
          },
          "incident_type": {
            "title": "Incident Type",
            "type": "string"
          },
          "signature_count": {
            "title": "Signature Count",
            "type": "integer"
          },
          "tenant_count": {
            "title": "Tenant Count",
            "type": "integer"
          }
        },
        "required": [
          "incident_type",
          "signature_count",
          "tenant_count",
          "avg_confidence"
        ],
        "title": "IncidentTypeBreakdownOut",
        "type": "object"
      },
...
      "IndicatorOverlapEntryOut": {
        "properties": {
          "first_observed_at": {
            "format": "date-time",
            "title": "First Observed At",
            "type": "string"
          },
          "incident_types": {
            "items": {
              "type": "string"
            },
            "title": "Incident Types",
            "type": "array"
          },
          "indicator_hash": {
            "title": "Indicator Hash",
            "type": "string"
          },
          "last_observed_at": {
            "format": "date-time",
            "title": "Last Observed At",
            "type": "string"
          },
          "signature_count": {
            "title": "Signature Count",
            "type": "integer"
          },
          "tenant_count": {
            "title": "Tenant Count",
            "type": "integer"
          }
        },
        "required": [
          "indicator_hash",
          "signature_count",
          "tenant_count",
          "incident_types",
          "first_observed_at",
          "last_observed_at"
        ],
        "title": "IndicatorOverlapEntryOut",
        "type": "object"
      },
      "IndicatorOverlapResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IndicatorOverlapEntryOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "IndicatorOverlapResponse",
        "type": "object"
      },
...
      "OverlapBucketOut": {
        "description": "`bucket` is one of `\"1\"`, `\"2\"`, `\"3+\"` \u2014 the number of distinct tenants that have\nseen a given indicator signature. The `\"2\"`/`\"3+\"` buckets together are the cross-tenant\nsignal itself; `\"1\"` is the (expected) majority of indicators seen by only one tenant.",
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "indicator_count": {
            "title": "Indicator Count",
            "type": "integer"
          }
        },
        "required": [
          "bucket",
          "indicator_count"
        ],
        "title": "OverlapBucketOut",
        "type": "object"
      },
      "OverlapDistributionResponse": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/OverlapBucketOut"
            },
            "title": "Buckets",
            "type": "array"
          },
          "total_indicators": {
            "title": "Total Indicators",
            "type": "integer"
          }
        },
        "required": [
          "total_indicators",
          "buckets"
        ],
        "title": "OverlapDistributionResponse",
        "type": "object"
      },
...
      "TechniquePrevalenceEntryOut": {
        "description": "One row per allowlisted technique (`data/kb/mitre/allowlist.yml`) \u2014 always all 13,\nincluding ones with `tenant_count == 0`, never a fabricated id.",
        "properties": {
          "signature_count": {
            "title": "Signature Count",
            "type": "integer"
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          },
          "technique_name": {
            "title": "Technique Name",
            "type": "string"
          },
          "tenant_count": {
            "title": "Tenant Count",
            "type": "integer"
          }
        },
        "required": [
          "technique_id",
          "technique_name",
          "tenant_count",
          "signature_count"
        ],
        "title": "TechniquePrevalenceEntryOut",
        "type": "object"
      },
      "TechniquePrevalenceResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TechniquePrevalenceEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total_tenants_with_signatures": {
            "title": "Total Tenants With Signatures",
            "type": "integer"
          }
        },
        "required": [
          "total_tenants_with_signatures",
          "items"
        ],
        "title": "TechniquePrevalenceResponse",
        "type": "object"
      },
...
      "Tier2OverviewResponse": {
        "properties": {
          "by_incident_type": {
            "items": {
              "$ref": "#/components/schemas/IncidentTypeBreakdownOut"
            },
            "title": "By Incident Type",
            "type": "array"
          },
          "total_overlapping_indicators": {
            "title": "Total Overlapping Indicators",
            "type": "integer"
          },
          "total_signatures": {
            "title": "Total Signatures",
            "type": "integer"
          },
          "total_tenants": {
            "title": "Total Tenants",
            "type": "integer"
          }
        },
        "required": [
          "total_signatures",
          "total_tenants",
          "total_overlapping_indicators",
          "by_incident_type"
        ],
        "title": "Tier2OverviewResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Layered detection funnel over security telemetry: OCSF normalization \u2192 rules \u2192 signal processing \u2192 entity-window ML \u2192 sequence models \u2192 graph correlation \u2192 agentic triage.",
    "title": "Tenex SOC Analyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tier2/first-seen": {
      "get": {
        "description": "Chart 4: for indicators seen by `min_tenants` or more tenants, when each tenant first\nobserved it \u2014 the early-warning story (\"tenant A on day 1, tenant B on day 4\").",
        "operationId": "get_first_seen_propagation_api_tier2_first_seen_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_tenants",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 1000,
              "minimum": 2,
              "title": "Min Tenants",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstSeenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get First Seen Propagation",
        "tags": [
          "tier2"
        ]
      }
    },
    "/api/tier2/indicator-overlap": {
      "get": {
        "operationId": "get_indicator_overlap_api_tier2_indicator_overlap_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_tenants",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 1000,
              "minimum": 2,
              "title": "Min Tenants",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicatorOverlapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Indicator Overlap",
        "tags": [
          "tier2"
        ]
      }
    },
    "/api/tier2/overlap-distribution": {
      "get": {
        "description": "Chart 1: for every indicator signature, how many distinct tenants have seen it,\nbucketed into 1 / 2 / 3+. The 2+ buckets are the cross-tenant signal itself.",
        "operationId": "get_overlap_distribution_api_tier2_overlap_distribution_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverlapDistributionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overlap Distribution",
        "tags": [
          "tier2"
        ]
      }
    },
    "/api/tier2/overview": {
      "get": {
        "operationId": "get_tier2_overview_api_tier2_overview_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tier2OverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tier2 Overview",
        "tags": [
          "tier2"
        ]
      }
    },
    "/api/tier2/stream": {
      "get": {
        "description": "SSE: one frame each time a pipeline run contributes to Tier 2.\n\nSubscribes to the single fleet-wide `TIER2_CHANNEL` rather than a per-analysis channel,\nbecause what changes for a Tier 2 viewer is the cross-tenant aggregate, not one run. The\nframe carries only counts \u2014 the client re-fetches, since every panel is a `GROUP BY` over the\nwhole store and a delta could not be applied to it correctly.\n\nAuthenticated like every other Tier 2 route, and like them not tenant-filtered: there is no\ntenant to filter by, since the store holds only `tenant_hash`. A subscriber learns that the\nfleet-wide picture moved, never whose data moved it.",
        "operationId": "stream_tier2_updates_api_tier2_stream_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Tier2 Updates",
        "tags": [
          "tier2"
        ]
      }
    },
    "/api/tier2/technique-prevalence": {
      "get": {
        "description": "Chart 2: which of the 13 proxy-observable ATT&CK techniques (docs/13 M14,\n`data/kb/mitre/allowlist.yml`) appear in how many tenants \u2014 every allowlisted technique\nis returned, including ones observed in zero tenants so far, never a fabricated id.",
        "operationId": "get_technique_prevalence_api_tier2_technique_prevalence_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "tenex_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenex Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechniquePrevalenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Technique Prevalence",
        "tags": [
          "tier2"
        ]
      }
    },
...
  }
}